chore: refactor get-songs
