+++ src/types.ts
@@ -13,5 +13,7 @@ export const TYPES = {
   Services: {
     AddQueryToQueue: Symbol('AddQueryToQueue'),
     GetSongs: Symbol('GetSongs'),
+    YoutubeAPI: Symbol('YoutubeAPI'),
+    SpotifyAPI: Symbol('SpotifyAPI'),
   },
 };
